// projects/series-movies/src/app/models/enum/settings-menu.enum.ts
export enum SettingsMenu {
  SEPARATOR = "SEPARATOR",

  KEYMAP = "SETTINGS.KEYMAP",
  LANGUAGE_AND_REGION = "SETTINGS.LANGUAGE_AND_REGION",

  DB = "SETTINGS.DB",
  IMPORT = "SETTINGS.IMPORT",
  DOWNLOAD = "SETTINGS.DOWNLOAD",
  REQUESTS = "SETTINGS.REQUESTS",
  TESTING = "SETTINGS.TESTING",

  ABOUT = "SETTINGS.ABOUT",
  HELP = "SETTINGS.HELP",
}

export enum SettingsMenuKey {
  KEYMAP = "keymap",
  LANGUAGE_AND_REGION = "language",

  DB = "db",
  IMPORT = "import",
  DOWNLOAD = "download",
  REQUESTS = "requests",
  TESTING = "testing",

  ABOUT = "about",
  HELP = "help",
}